# docs/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flaps HTTP API",
    "version": "0.1.0",
    "description": "Admin CRUD, SDK key management, SDK identity, OFREP evaluation and ruleset sync surface exposed by flaps-server. Generated by hand from build_router in crates/flaps-server/src/lib.rs (main@b4e15b1); see docs/spec/api-v1.md for the accompanying guide (authentication model, SSE contract, ETag semantics, error format)."
  },
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session bearer token minted by POST /login. Required by every /projects/** admin route (project, environment, flag, segment, flag-env-config and SDK key management)."
      },
      "sdkKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "SDK key bearer token. GET /sdk/whoami and the OFREP evaluation routes accept both server-kind and client-kind keys. The /sync/v1/* routes require a server-kind key; a client-kind key on those routes receives 403 Forbidden (see api-v1.md for the server-vs-client distinction, which this single security scheme does not model)."
      }
    },
    "parameters": {
...
      "IfNoneMatchHeader": {
        "name": "If-None-Match",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Optional ETag for conditional GET. An exact match short-circuits to 304 Not Modified with no body."
      }
    },
    "headers": {
      "ETagHeader": {
        "description": "Strong ETag: hex SHA-256 of the canonical (key-sorted) JSON serialization of the resource.",
        "schema": { "type": "string" }
      },
      "XFlapsVersionHeader": {
        "description": "Monotone version counter of the compiled ruleset.",
        "schema": { "type": "integer", "format": "int64" }
      },
...
      "RetryAfterHeader": {
        "description": "Suggested wait time in seconds before retrying, present on 429 responses.",
        "schema": { "type": "integer", "format": "int64" }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid authentication credentials.",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
      "ForbiddenServerKeyOnly": {
        "description": "The endpoint requires a server-kind SDK key; a client-kind key was supplied.",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
      "NotFound": {
        "description": "The addressed resource does not exist.",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded.",
        "headers": {
          "Retry-After": { "$ref": "#/components/headers/RetryAfterHeader" }
        },
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      },
      "InternalError": {
        "description": "Unexpected internal error (e.g. a persistence failure).",
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem+json error body.",
        "properties": {
          "type": { "type": "string", "format": "uri", "description": "Stable problem type URI, e.g. https://flaps.dev/problems/not-found." },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" }
        },
        "required": ["type", "title", "status", "detail"]
      },
...
      "RulesetDocument": {
        "type": "object",
        "description": "The compiled flagd-compatible ruleset document produced by flaps-compiler. Its internal schema is opaque here; it is generated, not hand-authored, and out of scope for this contract.",
        "additionalProperties": true
      },
      "EventPayload": {
        "type": "object",
        "description": "SSE frame payload on GET /sync/v1/events. Intentionally minimal: never carries flag data, only the notification that a new version exists.",
        "properties": {
          "environment": { "type": "string" },
          "version": { "type": "integer", "format": "int64" }
        },
        "required": ["environment", "version"]
      }
    }
  },
  "paths": {
...
    "/sync/v1/ruleset": {
      "get": {
        "summary": "Download the compiled ruleset for the SDK key's scope",
        "description": "Server-key only: a client-kind key receives 403. Supports If-None-Match/304 based on the ruleset content hash.",
        "operationId": "getRuleset",
        "security": [{ "sdkKey": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IfNoneMatchHeader" }],
        "responses": {
          "200": {
            "description": "The compiled ruleset document.",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETagHeader" },
              "X-Flaps-Version": { "$ref": "#/components/headers/XFlapsVersionHeader" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RulesetDocument" } } }
          },
          "304": { "description": "Not modified: the client's If-None-Match matches the current content hash. No body." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/ForbiddenServerKeyOnly" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/sync/v1/events": {
      "get": {
        "summary": "Server-sent events stream of ruleset change notifications",
        "description": "Server-key only: a client-kind key receives 403. Bounded by a separate concurrency quota (global and per-key) that rejects excess concurrent subscriptions with 429, distinct from the ordinary request rate limiter. See api-v1.md for the full notify-then-fetch SSE contract and the concurrency quota; text/event-stream is not modeled in detail by OpenAPI.",
        "operationId": "getEvents",
        "security": [{ "sdkKey": [] }],
        "responses": {
          "200": {
            "description": "SSE stream opened. Each frame carries a JSON-encoded EventPayload.",
            "content": { "text/event-stream": { "schema": { "$ref": "#/components/schemas/EventPayload" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/ForbiddenServerKeyOnly" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  }
}